requirements.txt
google-genai>=1.0.0
pinecone>=5.0.0
python-dotenv>=1.0.0
rich>=13.0.0
flask>=3.0.0
pymupdf>=1.24.0
